+++ src/app/routers/advisory_policy_evaluation_evidence.py
@@ -24,24 +24,6 @@ async def get_policy_evaluation_lineage(
 
 
 @router.get(
-    "/advisory-policy-evaluations/{evaluation_id}/sign-off-package",
-    response_model=AdvisoryPolicyEnvelopeResponse,
-    summary="Get Advisory Policy Sign-off Package",
-    description=(
-        "Returns the source-owned sign-off package from lotus-advise, including "
-        "maker-checker and client-ready blockers when supplied by Advise."
-    ),
-)
-async def get_policy_sign_off_package(
-    evaluation_id: str = POLICY_EVALUATION_PATH,
-) -> AdvisoryPolicyEnvelopeResponse:
-    return await advisory_policy_service().get_policy_sign_off_package(
-        evaluation_id=evaluation_id,
-        correlation_id=correlation_id_var.get(),
-    )
-
-
-@router.get(
     "/advisory-policy-evaluations/{evaluation_id}/workflow",
     response_model=AdvisoryPolicyEnvelopeResponse,
     summary="Get Advisory Policy Workflow",
